packages/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
